Move persistent gesture state to Document, add DocumentUserGestureToken
Non-test gesture tokens are now created as a DocumentUserGestureToken, which take a Document* (which can be null, but such uses are limited). When it is created, it sets the m_hasReceivedUserGesture flag on Document.

Remove UserGestureIndicator::s_processedUserGestureSinceLoad, as Document::m_hasReceivedUserGesture is very similar and not process-wide.

BUG=624061

Review-Url: https://codereview.chromium.org/2408333004
Cr-Original-Commit-Position: refs/heads/master@{#427481}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 26ce312cc8b3f53dc76df85e45c17bba93685fc3