Fetch retry summary and use it to determine what tests to rebaseline.
This CL changes RebaselineCL._tests_to_rebaseline so that after it fetches the layout test results and extracts a list of failing tests to rebaseline, it also tries to fetch the retry summary from Google Storage; if this is available, then it filters the tests to rebaseline to only include tests that failed with the patch but not without.

To support this, this CL also adds a method in the BuildBot and BuildBotMock classes.

BUG=642980

Review-Url: https://codereview.chromium.org/2446463002
Cr-Original-Commit-Position: refs/heads/master@{#427536}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f9833f761d7bd907e91fcfbc669aaa5235a5a471