CSP: 'violated-directive' should match 'effective-directive'.
In [1], mikispag@ noted that we're currently returning the entire
text of the violated directive, including its source list, etc. This
information is replicated in the report via the 'original-policy'
property, and CSP3 defines both 'violated-directive' and
'effective-directive' to contain the same data.

This patch changes our implementation in line with the spec.

[1]: https://github.com/w3c/web-platform-tests/pull/4020

R=jochen@chromium.org

Review-Url: https://codereview.chromium.org/2447003002
Cr-Original-Commit-Position: refs/heads/master@{#427634}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0e64be9cf335ee3bea7c989702c5a9a0934af037