Fix payments layout test timeout.
A payments layout test was timing out due to not being able to find the
mojo file that it was mocking. That file has moved from
third_party/WebKit to components/payments. Updating the location of the
mojo file fixes the timeout.

The patch also removes a stale reference to this mojo file from DEPS.

BUG=659391

Review-Url: https://codereview.chromium.org/2456483003
Cr-Original-Commit-Position: refs/heads/master@{#427998}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a02160012bad1d6486ddcc17966fa585fd585cbb