child-src and frame-src CSP are not applicable when navigating a new window.
Shift-clicking an anchor/link will open the link in a new window
(similarily ctrl-clicking or middle-clicking will open the link
in a new background tab).  In this scenario child-src and frame-src
Content Security Policies of the parent frame are not applicable
(because we are not navigating the frame containing the anchor,
but instead we are navigating a brand new frame in a new window).

BUG=658701

Review-Url: https://codereview.chromium.org/2453093003
Cr-Original-Commit-Position: refs/heads/master@{#428069}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d7e5f244d54d0a0c8615e4ff216f906851e9fb64