[root layer scrolling] Update LayoutGeometryMapTest for RLS.
LayoutGeometryMap and LayoutObject::mapLocalToAncestor used to stop
short of applying the frame-level scroll offset.  With root layer
scrolling enabled, the frame-level scroll offset is applied by the
generic offsetFromContainer() implementation, so it will be included
in coordinates mapped up to the top level.

This patch updates test expectations to allow for the frame-level
scroll offset to be included when root layer scrolling is enabled.

BUG=490942
R=skobes@chromium.org,chrishtr@chromium.org

Review-Url: https://codereview.chromium.org/2417103002
Cr-Original-Commit-Position: refs/heads/master@{#428113}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1b4ff08e37e5a20369d085cb69e20a706fc8ca7e