Third-party iframe logging: Separate out AboveAndLeft from Above
Early data shows that iframes above the page, which includes iframes
that are simultaneously above AND left of the page, are 25x more common
than iframes only to the left of the page. It's possible that
AboveAndLeft are most common, which could explain the discrepancy.

BUG=635105

Review-Url: https://codereview.chromium.org/2450293003
Cr-Original-Commit-Position: refs/heads/master@{#428278}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e10fbfce772e735f8b5fbedb94a67bff8e8f8e91