Use setSequentialFocusNavigationStartingPoint in ChromeVox
Improves the interaction between ChromeVox navigation and input focus.
If you navigate to a focusable node, focuses it. If you navigate to a
descendant of a focusable node, focuses it too (checks the common
ancestors of the start and end of the selection). If not, sets the
sequential focus navigation start point to the start of the selection
so that subsequent focus using the Tab key is relative to that point.

Includes two new tests.

BUG=655272
TESTED=Visit Wikipedia, press Cvox+H to navigate to heading, then Tab to go to the next link after that heading. Then navigate by line with Search+Up/Down and confirm that it focuses links as you pass over them.
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/2447773002
Cr-Original-Commit-Position: refs/heads/master@{#428313}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fb6b8fd66c69ba0460fc38bbe3a256876665d7cf