INPUT element: Updating 'step' attribute should adjust the last value.
Bug:
If
 - 'step' attribute is updated,
 - we adjust the value, and
 - the value is not dirty,
we adjusted the default value instead of the last value.

Implementation:
We didn't store the last value if the value is not dirty, and compute the value
on the fly.
With this CL, we always store the current value to
HTMLInputElement::m_valueIfDirty though we stored only a dirty value.
 * When |value| attribute is updated, we update m_valueIfDirty as well.
 * Remove special handling of null m_valueIfDirty
  Because setValue(null0string) doesn't clear hasDirtyValue() flag, this CL
  introduces setNonDirtyValue() like HTMLTextAreaElement::setNonDirtyValue().
 * Make HTMLInputElement::value() clearer.

BUG=633074

Review-Url: https://codereview.chromium.org/2463693002
Cr-Original-Commit-Position: refs/heads/master@{#428663}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9763e857c8822745fb03e460e1bafdb638b95d71