The dragleave and dragend events should not be cancelable.
Per the HTML spec, the dragleave (fired at targets) and dragend (fired
at the source) events should not be cancelable. We already don't do
anything if preventDefault() is called, but we should report the
correct properties on the events when inspected from script.

BUG=272248

Review-Url: https://codereview.chromium.org/2465433003
Cr-Original-Commit-Position: refs/heads/master@{#428779}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6b83fd79210649e6408e05b15b5c4969c82fe8e7