FaceDetection: add support for <video> input
This CL adds support for <video> as input to detect().
Following canvas approach, what is done is to snap the
current video frame and use for the detection process.

- Some code is refactored in FaceDetector.cpp since,
from a c++ perspective, an ImageBitmap and an
HTMLVideoElement behave similarly.

- some LayoutTests code is refactored since, from the
JS/DOM perspective, an HMTLImageElement and an
HTMLVideoElement behave similarly.

BUG=659138
TEST=LayoutTests.

Review-Url: https://codereview.chromium.org/2455973007
Cr-Original-Commit-Position: refs/heads/master@{#429016}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 18e8e2880dea2fa45b5edf85a418c12f53547f01