Fix incorrect usage of std::is_heap in blink::scheduler::IntrusiveHeap.
IntrusiveHeap uses std::is_heap in a DCHECK, but it uses a comparator
function that isn't irreflexive (the comparator is like <, but it should
be like <=). This CL makes the (super simple fix) and adds the necessary
plumbing for IntrusiveHeap's comparator to be unit-tested.

BUG=661080

Review-Url: https://codereview.chromium.org/2469653002
Cr-Original-Commit-Position: refs/heads/master@{#429026}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b32830fb14c0a6c36b2735089fdb3a2ce26106e0