Add a UA style, "display: contents", for <slot>
The spec is: https://html.spec.whatwg.org/#flow-content-3

This rule is guarded by the same runtime flag for "display: contents".
Unless the flag is enabled, the rule can not be effective because the parser does
not consider it a valid rule.

This change only affects slots in non-shadow trees since Blink excludes slots in
shadow trees from a flat tree.

Since "display: contents" is not well supported yet, the children of a slot in
non-shadow trees would not be displayed.

BUG=657748,660265

Review-Url: https://codereview.chromium.org/2466243002
Cr-Original-Commit-Position: refs/heads/master@{#429198}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d582fd6159457a8a9ff4f2064ce45d8bc4626c72