Convert editing/spelling/spellcheck-attribute-settings-default.html with spellcheck_test
This patch converts the layout test with spellcheck_test as a
preparation for implementing idle time spellchecker.

This patch also fixes a bug in the last test case of the original file,
where element with empty-string spellcheck attribute should be spellchecked
when the global default value is false, while the original test case asserts
that the element is not checked.

BUG=517298

Review-Url: https://codereview.chromium.org/2463323002
Cr-Original-Commit-Position: refs/heads/master@{#429219}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6ee9acea67b949b32cf2b1eff8ff6a7a17d90774