Make the baseline of LayoutTextControl consistent with or without text
This patch fixes LayoutTextControl::firstLineBoxBaseline() to compute
using the primary font when the text is missing.

Before the fix, it computes to -1 when there is no lineboxes. This
resulted in different baseline position when <input> or <textarea> has
"display: block" and text is empty.

BUG=490453

Review-Url: https://codereview.chromium.org/2473613002
Cr-Original-Commit-Position: refs/heads/master@{#429246}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a7c29f036fe5a14424096c7b021285b7faa1799f