Fix UAF when sibling and ancestor scrollers share an anchor node.
It's not safe to reset the IsScrollAnchorObject bit "unconditionally", even when
we are clearing all ancestor ScrollAnchor objects from a scroller, because the
anchor node may still be referenced by a scroller that is not an ancestor of the
one being cleared.

Sadly this means that ScrollAnchor::clear() may be O(n^2) in the depth of the
PaintLayer tree, but I can't think of a good way around this.

BUG=659890

Review-Url: https://codereview.chromium.org/2471653003
Cr-Original-Commit-Position: refs/heads/master@{#429281}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0de38f1b237078e0c9cb46491de2df25a0f37446