Move debug_devtools to a buildflag_header.
Previously this was in the global features.gni file, and files did an ad-hoc
reference to the debug_devtools config if they needed the define. This is
fragile and will not scale as additional flags are added.

This patch introduces a Blink public features buildflag_header (in the
future I expect we'll also have a Blink-internal one in Source/). The
existing users of debug_devtools are updated to use this.

BUG=

Review-Url: https://codereview.chromium.org/2458573004
Cr-Original-Commit-Position: refs/heads/master@{#429362}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cb4ab75ac5e94b7857b70a9d6f73eefe9db47a7f