Remove ScriptState::Scope in V8ScriptValueSerializer.
This avoids a DCHECK failure when the context is going away, and the context
will already be entered when we call into V8 anyhow (besides which, in practice
it is also the current context).

This corresponds to what ScriptValueSerializer already does (or doesn't do).

BUG=148757

Review-Url: https://codereview.chromium.org/2462093002
Cr-Original-Commit-Position: refs/heads/master@{#429407}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 997b2ce593881d7cfddaa08df115188feef3be6f